compost(api): split utils.py god-module into focused kernel_* router families
utils.py had accreted to 2625 lines / 23 routes as transmuted kernel-served
endpoints were added one at a time, tripping the maintainability audit's
very_large flag (>=1000 lines) and large flag (>=600) — it sat in BOTH lists.

Pure structural MOVE, zero behavior change: the 22 transmuted endpoints split
by family into six focused modules, each well under the 600-line large
threshold. All share one /utils APIRouter defined in the new kernel_shared.py
(plus the four query/result coercion helpers), so every route keeps its exact
/api/utils/... path and app.main still includes one router with one call.

  kernel_shared.py    — the shared router + logger + parse/coerce helpers
  kernel_nodeid.py    — coherence_weight, nodeid_distance, nodeid_compatibility
  kernel_scoring.py   — weighted_average, simpson_diversity, idea_score,
                        marginal_cc_return, idea_marginal_from_record, softmax_weights
  kernel_breath.py    — breath_balance, shannon_entropy, coherence_summary_score
  kernel_grounding.py — cost_vector, value_vector, grounded_roi,
                        idea_grounding_summary, idea_grounded_cost_sum
  kernel_grounded_cv.py — grounded_cost, grounded_value
  kernel_matching.py  — concept_match_score, tag_match_score, worldview_alignment

utils.py stays as the slim aggregator: it imports the family modules (running
their @router.get decorators), keeps kernel_status, and re-exports the nine
_py fallbacks the route tests import so historical
`from app.routers.utils import <name>_py` paths stay stable.

Every function body moved verbatim (45/45 functions byte-identical to the
pre-split source); no recipe value or route path changed.

Proof:
- maintainability audit: very_large 16 -> 15 (= baseline, the regression I
  introduced cleared); utils.py no longer large OR very_large. (Residual
  large 44 > baseline 43 is pre-existing main drift in unrelated service
  modules, not this accretion.)
- 22/22 kernel-served routes trace value-parity (recipes unchanged).
- full API suite: 2446 passed, 3 skipped; 93 utils route tests green.
- app imports clean; all 23 /api/utils/... paths resolve unchanged.

Edges in the same breath: routers INDEX.md + MANIFEST.md regenerated.
